pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pglease"
version = "0.1.0"
description = "Lightweight distributed task coordination library"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Raul Guiu Gallardo"}
]
keywords = ["distributed", "coordination", "locking", "singleton", "kubernetes"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.9.0",  # linter + formatter (replaces black + isort + flake8)
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/pglease"
Repository = "https://github.com/yourusername/pglease"

[tool.setuptools.packages.find]
where = ["src"]

# ---------------------------------------------------------------------------
# Ruff — linter + formatter
# Replaces: black, isort, flake8, pyupgrade
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes (undefined names, unused imports, …)
    "I",    # isort (import ordering)
    "UP",   # pyupgrade (modernise Python syntax)
    "B",    # flake8-bugbear (likely bugs & design issues)
    "SIM",  # flake8-simplify
    "N",    # pep8-naming
    "C4",   # flake8-comprehensions
    "RET",  # flake8-return
]
ignore = [
    "E501",    # line-too-long — enforced by the formatter instead
    "B008",    # do-not-call-function-in-default-argument
    "RET505",  # superfluous-else-return — style preference
    "SIM108",  # ternary-instead-of-if-else — readability preference
]

[tool.ruff.lint.per-file-ignores]
# Tests use assert statements and occasionally pytest.raises(Exception)
"tests/**" = ["B017", "F811"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
